+++ app/controllers/staff/reports_state_controller.rb
@@ -60,7 +60,6 @@ class Staff::ReportsStateController < Staff::BaseController
     unless report.state == state
       authorize report, policy_action + "?"
       report.update!(state: state)
-      find_or_create_new_report(organisation_id: report.organisation.id, fund_id: report.fund.id) if state == "approved"
 
       Report::SendStateChangeEmails.new(report).send!
     end
@@ -71,9 +70,5 @@ class Staff::ReportsStateController < Staff::BaseController
 
   private def report
     @report ||= Report.lock.find(params[:report_id])
-  end
-
-  private def find_or_create_new_report(organisation_id:, fund_id:)
-    Report.where.not(state: :approved).find_or_create_by!(organisation_id: organisation_id, fund_id: fund_id)
   end
 end