+++ app/services/create_adjustment.rb
@@ -28,7 +28,8 @@ class CreateAdjustment
     Adjustment.new(adjustment_attrs).tap do |adjustment|
       adjustment.build_comment(
         body: attributes.fetch(:comment),
-        commentable: adjustment
+        commentable: adjustment,
+        report: report,
       )
       adjustment.build_detail(
         user: user,
@@ -69,8 +70,8 @@ class CreateAdjustment
   end
 
   def bail_if_report_not_acceptable
-    if report.state != "active"
-      msg = "Report ##{report.id} is not in the active state"
+    unless Report::EDITABLE_STATES.include?(report.state)
+      msg = "Report ##{report.id} is not in an editable state"
     end
     unless valid_reports_for_activity.include?(report)
       msg = "Report ##{report.id} is not associated with Activity ##{activity.id}"