+++ db/migrate/20211014155109_remove_legacy_comments.rb
@@ -1,0 +1,9 @@
+class RemoveLegacyComments < ActiveRecord::Migration[6.1]
+  def up
+    drop_table :legacy_comments
+  end
+
+  def down
+    raise ActiveRecord::IrreversibleMigration
+  end
+end