+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_10_12_141009) do
+ActiveRecord::Schema.define(version: 2021_10_14_155109) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -222,18 +222,6 @@ ActiveRecord::Schema.define(version: 2021_10_12_141009) do
     t.index ["destination_id"], name: "index_incoming_transfers_on_destination_id"
     t.index ["report_id"], name: "index_incoming_transfers_on_report_id"
     t.index ["source_id"], name: "index_incoming_transfers_on_source_id"
-  end
-
-  create_table "legacy_comments", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
-    t.text "comment"
-    t.uuid "owner_id"
-    t.uuid "activity_id"
-    t.uuid "report_id"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
-    t.index ["activity_id"], name: "index_legacy_comments_on_activity_id"
-    t.index ["owner_id"], name: "index_legacy_comments_on_owner_id"
-    t.index ["report_id"], name: "index_legacy_comments_on_report_id"
   end
 
   create_table "matched_efforts", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|